+++ src/utils/index.ts
@@ -6,3 +6,4 @@ export * from './autoconnect'
 export * from './price'
 export * from './validation'
 export * from './flagCodes'
+export * from './deferred'
